Extract enums for menus
